requirements.txt
biosimulators_utils[logging] >= 0.1.53
cobra
lxml
numpy
python_libsbml
